Add addExtra method
This method differs from setExtra, as it is expected to only add properties to extra,
not replace whole extra object.
If there are new properties (that are not in extra), they will be added.
If there are existing properties (that are in extra), they will be overwritten by new ones.