Bring passing `namespace` to stash-it itself
This is a more wider subject.

As there is just one adapter instance being passed in order to create
the cache object, namespace can be brought up to this level of abstraction,
making it less complicated on the adapter level.

Initial idea, with namespace being set in adapter, was that adapter would
know what characters can be used to construct the namespace so that it
(namespace) can be used for the storage given adapter uses.

Should there be a problem (as described above), it will be just up to
the user of stash-it to change the namespace string (which is easy
enough).